fix duplicate xlm_print_title issue
git-svn-id: http://npoi.googlecode.com/svn/trunk@519 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ ooxml/XSSF/UserModel/XSSFCreationHelper.cs
@@ -45,7 +45,7 @@ namespace NPOI.XSSF.UserModel
 
         public IHyperlink CreateHyperlink(HyperlinkType type)
         {
-            return new XSSFHyperlink((int)type);
+            return new XSSFHyperlink(type);
         }
 
         /**
